+++ src/main/scala/advxml/core/transform/actions/AttributeData.scala
@@ -1,0 +1,5 @@
+package advxml.core.transform.actions
+
+import scala.xml.Text
+
+case class AttributeData(key: String, value: Text)
