+++ src/main/scala/advxml/core/transform/actions/XmlPredicate.scala
@@ -1,0 +1,10 @@
+package advxml.core.transform.actions
+
+import scala.xml.NodeSeq
+
+object XmlPredicate {
+
+  type XmlPredicate = NodeSeq => Boolean
+
+  def apply(f: NodeSeq => Boolean): XmlPredicate = f(_)
+}
