+++ src/main/scala/advxml/core/transform/actions/XmlZoom.scala
@@ -1,0 +1,8 @@
+package advxml.core.transform.actions
+
+import scala.xml.NodeSeq
+
+object XmlZoom {
+  type XmlZoom = NodeSeq => NodeSeq
+  def apply(f: NodeSeq => NodeSeq): XmlZoom = f(_)
+}
