--- src/main/scala/advxml/core/transformation/actions/AttributeData.scala
@@ -1,5 +1,0 @@
-package advxml.core.transformation.actions
-
-import scala.xml.Text
-
-case class AttributeData(key: String, value: Text)
