--- src/main/scala/advxml/core/transformation/actions/XmlZoom.scala
@@ -1,8 +1,0 @@
-package advxml.core.transformation.actions
-
-import scala.xml.NodeSeq
-
-object XmlZoom {
-  type XmlZoom = NodeSeq => NodeSeq
-  def apply(f: NodeSeq => NodeSeq): XmlZoom = f(_)
-}
