+++ src/main/scala/advxml/instances/AllInstances.scala
@@ -1,5 +1,3 @@
 package advxml.instances
 
-import advxml.instances.transformation.XmlTransformerInstances
-
 private[advxml] trait AllInstances extends XmlTransformerInstances with ConvertersInstances with ValidationInstance
