--- src/main/scala/advxml/instances/transformation/XmlTransformerInstances.scala
@@ -1,3 +1,0 @@
-package advxml.instances.transformation
-
-trait XmlTransformerInstances extends AllXmlModifierInstances with XmlPredicateInstances
