+++ src/test/scala/advxml/syntax/XmlConverterSyntaxTest.scala
@@ -1,7 +1,7 @@
 package advxml.syntax
 
-import advxml.core.converters.{ModelToXml, XmlToModel}
-import advxml.core.validation.ValidatedEx
+import advxml.core.convert.{ModelToXml, XmlToModel}
+import advxml.core.validate.ValidatedEx
 import cats.data.Kleisli
 import cats.data.Validated.Valid
 import org.scalatest.FunSuite
@@ -28,7 +28,7 @@ class XmlConverterSyntaxTest extends FunSuite {
       (
         (x \@! "Name"),
         (x \@! "Surname"),
-        (x \@? "Age").map(_.map(_.toInt))
+        (x \@? "Age").deepMap(_.toInt)
       ).mapN(Person)
     })
